fix authentication and start review

+++ src/main/java/ru/saynurdinov/moviefan/repository/ActorRepository.java
@@ -3,6 +3,8 @@ package ru.saynurdinov.moviefan.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.saynurdinov.moviefan.model.Actor;
 
-public interface ActorRepository extends JpaRepository<Actor, Integer> {
+import java.util.List;
+
+public interface ActorRepository extends JpaRepository<Actor, Long> {
 
 }
