+++ src/main/java/ru/saynurdinov/moviefan/repository/CollectionRepository.java
@@ -3,5 +3,5 @@ package ru.saynurdinov.moviefan.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.saynurdinov.moviefan.model.Collection;
 
-public interface CollectionRepository extends JpaRepository<Collection, Integer> {
+public interface CollectionRepository extends JpaRepository<Collection, Long> {
 }
